third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-valid-rgb-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['color'] = "rgba(calc(50% + (sign(1em - 10px) * 10%)) 400% -400% / 0.5)" should set the property value
  assert_equals: serialization should be canonical expected "rgb(calc(50% + (10% * sign(1em - 10px))) 255 0 / 0.5)" but got "rgb(calc(50% + (10% * sign(1em - 10px))) 4 0 / 0.5)"
[FAIL] e.style['color'] = "rgba(calc(50% + (sign(1em - 10px) * 10%)), 400%, -400%, 0.5)" should set the property value
  assert_equals: serialization should be canonical expected "rgb(calc(50% + (10% * sign(1em - 10px))) 255 0 / 0.5)" but got "rgb(calc(50% + (10% * sign(1em - 10px))) 4 0 / 0.5)"
Harness: the test ran to completion.
